add linq projection example

+++ src/MongoDb.Csharp.Samples/Models/User.cs
@@ -24,7 +24,7 @@ namespace MongoDb.Csharp.Samples.Models
         public string Website {get; set; }
         public CompanyCard Company {get; set; }
         public decimal Salary { get; set; }
-        public decimal MonthlyExpenses { get; set; }
+        public int MonthlyExpenses { get; set; }
         public List<string> FavoriteSports { get; set; }
         public string Profession { get; set; }
     }
